pyproject.toml
[tool.poetry]
name = "double-agent"
version = "0.1.0"
description = "Build a simple agent that is capable of hacking itself."
authors = ["James Levy <jamsalevy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.8"
openai = "^0.28.0"
pydantic = "^2.5.2"
click = "^8.1.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
